+++ internal/business/action/create_implicit_account_test.go
@@ -28,7 +28,7 @@ func TestUnmarshal(t *testing.T) {
 			)
 			assert.Equal(
 				t,
-				int64(10),
+				float64(10),
 				action.Balance,
 				"Assert balance",
 			)
